test(control-flow): isolate AERROR runtime metadata fixture
