+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+- 2026-08-21: Continued #2564 test-module refactoring by moving the focused
+  `AERROR()` runtime-metadata fixture into
+  `test_prg_engine_control_flow_aerror_runtime_metadata.cpp`. The existing
+  executable, invocation order, assertions, and runtime behavior remain
+  unchanged.
+
 - 2026-08-21: Continued #2564 test-module refactoring by moving the cursor/SQL
   locale-diagnostic fixtures into `test_prg_engine_control_flow_localized_errors.cpp`.
   The existing executable, invocation order, assertions, and runtime behavior
